refactor: rename `ProcessedResult` and move to `programming.base` module
