+++ unicon_backend/evaluator/tasks/programming/base.py
@@ -1,7 +1,6 @@
-from collections import Counter
 from functools import cached_property
 from logging import getLogger
-from typing import Any, Literal, Self
+from typing import Any, Literal, Self, cast
 
 from pydantic import BaseModel, RootModel, model_validator
 
@@ -11,10 +10,12 @@ from unicon_backend.evaluator.tasks.programming.security import mpi_sandbox
 from unicon_backend.evaluator.tasks.programming.steps import (
     ComputeGraph,
     InputStep,
+    OutputStep,
     StepSocket,
     StepType,
 )
-from unicon_backend.runner import ComputeContext, JobId, RunnerJob, RunnerProgram
+from unicon_backend.lib.common import CustomSQLModel
+from unicon_backend.runner import ComputeContext, JobId, ProgramResult, RunnerJob, RunnerProgram
 from unicon_backend.workers.publisher import task_publisher
 
 logger = getLogger(__name__)
@@ -25,16 +26,31 @@ USER_INPUT_STEP_ID: int = 0
 class Testcase(ComputeGraph):
     id: int
 
-    @cached_property
-    def socket_type_counts(self) -> Counter[str]:
-        return Counter(socket.type for socket in self.nodes)
-
     @model_validator(mode="after")
     def check_exactly_one_output_step(self) -> Self:
-        output_step_count = self.socket_type_counts[StepType.OUTPUT.value]
-        if output_step_count != 1:
-            raise ValueError(f"Expected exactly 1 output step, found {output_step_count}")
+        num_output_steps: int = len([node for node in self.nodes if node.type == StepType.OUTPUT])
+        if num_output_steps != 1:
+            raise ValueError(f"Expected exactly 1 output step, found {num_output_steps}")
         return self
+
+    @cached_property
+    def output_step(self) -> OutputStep:
+        return cast(OutputStep, next(node for node in self.nodes if node.type == StepType.OUTPUT))
+
+
+class SocketResult(CustomSQLModel):
+    """
+    This class is used to store whether the result of an output socket is right or wrong.
+    Note that whether or not to show this information (public) and other variables should be derived from data in Testcase.
+    """
+
+    id: str
+    value: Any
+    correct: bool
+
+
+class TestcaseResult(ProgramResult):
+    results: list[SocketResult] | None = None
 
 
 class RequiredInput(BaseModel):
